Implement veneer for ISA compat checks
