+++ Source/astcenc_entry.cpp
@@ -158,48 +158,6 @@ static astcenc_error validate_cpu_float()
 }
 
 /**
- * @brief Validate CPU ISA support meets the requirements of this build of the library.
- *
- * Each library build is statically compiled for a particular set of CPU ISA features, such as the
- * SIMD support or other ISA extensions such as POPCNT. This function checks that the host CPU
- * actually supports everything this build needs.
- *
- * @return Return @c ASTCENC_SUCCESS if validated, otherwise an error on failure.
- */
-static astcenc_error validate_cpu_isa()
-{
-	#if ASTCENC_SSE >= 41
-		if (!cpu_supports_sse41())
-		{
-			return ASTCENC_ERR_BAD_CPU_ISA;
-		}
-	#endif
-
-	#if ASTCENC_POPCNT >= 1
-		if (!cpu_supports_popcnt())
-		{
-			return ASTCENC_ERR_BAD_CPU_ISA;
-		}
-	#endif
-
-	#if ASTCENC_F16C >= 1
-		if (!cpu_supports_f16c())
-		{
-			return ASTCENC_ERR_BAD_CPU_ISA;
-		}
-	#endif
-
-	#if ASTCENC_AVX >= 2
-		if (!cpu_supports_avx2())
-		{
-			return ASTCENC_ERR_BAD_CPU_ISA;
-		}
-	#endif
-
-	return ASTCENC_SUCCESS;
-}
-
-/**
  * @brief Validate config profile.
  *
  * @param profile   The profile to check.
@@ -464,7 +422,7 @@ static astcenc_error validate_config(
 }
 
 /* See header for documentation. */
-astcenc_error astcenc_config_init(
+astcenc_error astcenc_config_init_actual(
 	astcenc_profile profile,
 	unsigned int block_x,
 	unsigned int block_y,
@@ -474,14 +432,6 @@ astcenc_error astcenc_config_init(
 	astcenc_config* configp
 ) {
 	astcenc_error status;
-
-	// Check basic library compatibility options here so they are checked early. Note, these checks
-	// are repeated in context_alloc for cases where callers use a manually defined config struct
-	status = validate_cpu_isa();
-	if (status != ASTCENC_SUCCESS)
-	{
-		return status;
-	}
 
 	status = validate_cpu_float();
 	if (status != ASTCENC_SUCCESS)
@@ -694,19 +644,13 @@ astcenc_error astcenc_config_init(
 }
 
 /* See header for documentation. */
-astcenc_error astcenc_context_alloc(
+astcenc_error astcenc_context_alloc_actual(
 	const astcenc_config* configp,
 	unsigned int thread_count,
 	astcenc_context** context
 ) {
 	astcenc_error status;
 	const astcenc_config& config = *configp;
-
-	status = validate_cpu_isa();
-	if (status != ASTCENC_SUCCESS)
-	{
-		return status;
-	}
 
 	status = validate_cpu_float();
 	if (status != ASTCENC_SUCCESS)
